Add files via upload
This directory presents a storytelling narrative on delayed deliveries. It begins with an exploratory analysis to identify the underlying problem. Subsequently, a database is built using Python and SQL to store the relevant data. Finally, the relationship between the variables is showcased, highlighting the urgent need for changes within the delivery sector due to the high volume of delayed orders.